fixing blog flies
