+++ controllers/index.js
@@ -2,6 +2,7 @@ const router = require('express').Router();
 
 const apiRoutes = require('./api');
 const homeRoutes = require('./homeRoutes');
+const blogRoutes = require('./api');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
