+++ views/blog.handlebars
@@ -1,4 +1,4 @@
-<h2>Blog</h2>
+{{!-- <h2>Blog</h2>
 
     <nav class="navbar navbar-dark bg-dark">
       <a class="navbar-brand" href="/">Note Taker </a>
@@ -26,16 +26,17 @@
       </div>
     </div>
     <script src="/public/js/blog.js"></script>
+ --}}
 
-{{!-- 
-
-    {{#each users as |user| }}
+{{#each notes as |note| }}
 <div class="row align-center mb-5">
   <div class="col-md-6">
-    <h2>{{{get_emoji}}} {{user.name}}</h2>
+    <h2>{{{get_emoji}}} {{blog.title}}</h2>
   </div>
   <div class="col-md-6">
-    <h3><a href="mailto:{{user.email}}">{{user.email}}</a></h3>
+    <h3>{{blog.text}}</a></h3>
   </div>
 </div>
-{{/each}} --}}
+{{/each}}
+
+  <script src="/public/js/blog.js"></script>
